+++ src/test/java/com/gonghak98/v2/audit/jpa/JpaGonghakRequirementRepositoryTest.java
@@ -1,10 +1,11 @@
-package com.gonghak98.v2.report.infrastructure.jpa;
+package com.gonghak98.v2.audit.jpa;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import com.gonghak98.v2.audit.infrastructure.jpa.JpaGonghakRequirementRepository;
 import com.gonghak98.v2.core.infrastructure.entity.DepartmentEntity;
 import com.gonghak98.v2.core.infrastructure.jpa.JpaDepartmentRepository;
-import com.gonghak98.v2.report.infrastructure.entity.GonghakRequirementEntity;
+import com.gonghak98.v2.audit.infrastructure.entity.GonghakRequirementEntity;
 import jakarta.persistence.EntityManager;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
